Move insert_column to sqlite.query_helpers
